refactor: setup패키지를 board패키지 내부로 이동
