+++ src/main/java/janggi/controller/JanggiController.java
@@ -4,9 +4,9 @@ import janggi.domain.GameManager;
 import janggi.domain.Position;
 import janggi.domain.board.Board;
 import janggi.domain.board.BoardGenerator;
+import janggi.domain.board.setup.ElephantFormation;
+import janggi.domain.board.setup.SetupStrategy;
 import janggi.domain.piece.Piece;
-import janggi.domain.setup.ElephantFormation;
-import janggi.domain.setup.SetupStrategy;
 import janggi.domain.team.BlueTeam;
 import janggi.domain.team.RedTeam;
 import janggi.domain.team.Team;
